Update 'index.html' and add 'style.css'
Added a new div class called, 'hero-card', to seperate the link tag and a div id for individual cards.

Implemented links into the a tags.

Create all styles for html to center and organize cards in the middle of the screen.

Added the hover pseudo-class to change the background-color, box-shadow, and text-decoration.